Moved files to their appropriate directories according to agreed file structure. (#20)
