+++ App.js
@@ -7,13 +7,13 @@ import SearchUserScreenController from "./components/search/SearchUserScreenCont
 import ChatHomeScreenController from "./components/chat/ChatHomeScreenController";
 import colorDefaults from "./theme/colorDefaults";
 import DirectMessageScreenController from "./components/chat/DirectMessageScreenController";
-import AuthContext from "./components/AuthContext";
+import AuthContext from "./context/AuthContext";
 import { auth } from "./firebase/firebaseConfig";
-import WelcomeScreen from "./Screens/WelcomeScreen";
-import SignInScreen from "./Screens/SignInScreen";
-import SignUpScreen from "./Screens/SignUpScreen";
-import ResetPassword from "./Screens/ResetPassword";
-import ChangePasswordScreen from "./Screens/ChangePasswordScreen";
+import WelcomeScreen from "./components/login/WelcomeScreen";
+import SignInScreen from "./components/login/SignInScreen";
+import SignUpScreen from "./components/login/SignUpScreen";
+import ResetPassword from "./components/login/ResetPassword";
+import ChangePasswordScreen from "./components/settings/ChangePasswordScreen";
 import StaffProfile from "./components/profile/staff/StaffProfile";
 import PatientProfile from "./components/profile/patient/PatientProfile";
 import Menu from "./components/mainmenu/Menu";
