+++ components/chat/DirectMessageScreenController.js
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect, useLayoutEffect, useState } from "react";
 import firestoreService from "../../firebase/firestoreService";
-import AuthContext from "../AuthContext";
+import AuthContext from "../../context/AuthContext";
 import Staff from "../../models/Staff";
 import DirectMessageScreenView from "./DirectMessageScreenView";
