+++ components/profile/staff/ProfileTab.js
@@ -10,7 +10,7 @@ import {
 import ProfileInformation from "../profilecomponents/ProfileInformation";
 import firestoreService from "../../../firebase/firestoreService";
 import { ScrollView } from "react-native-gesture-handler";
-import AuthContext from "../../AuthContext";
+import AuthContext from "../../../context/AuthContext";
 import EditEnableButton from "../profilecomponents/EditEnableButton";
 
 function ProfileTab({ user, setUser }) {
